refactor : featrue post 폴더명 변경
