+++ src/features/post/ui/PostDetailModal.tsx
