+++ src/features/post/ui/PostReactInfoTableCell.tsx
